Added support for floats. Changed int usage to int64. Added some more arithmatic commands. Added select/join

+++ src/secd/command/function.go
@@ -5,7 +5,7 @@ import (
 )
 
 type Function struct {
-	ArgCount int
+	ArgCount int64
 	Commands []Command
 	Environment *util.Stack
 }
